Make base django settings a .env file stored in SecretManager
To reduce the complexity of having to setup 5 secrets, this commit
changes the base django settings to be a file stored in secret manager,
pulled in the event that there is no local .env file (removing the
template .env file).

Shell changes made to suit.
Terraform changes made to suit, using a templatefile.

Also combines moving settings.py to unicodex/settings.py, and associated
file changes